# dsh/dsh.openapi.yaml
openapi: 3.1.0
info:
  title: BThwani DSH API
  version: 0.3.0
  description: API contract for client-facing store discovery, catalog management, and checkout flow.
  contact:
    name: BThwani DSH API Governance
    url: https://bthwani.internal/governance/dsh
servers:
  - url: https://api.bthwani.internal/dsh
    description: Production DSH API boundary
  - url: http://localhost:8080
    description: Local DSH backend runtime
...
paths:
...
  /stores/{id}/readiness-escalations:
    post:
      tags:
        - Field Readiness
      summary: Escalates incomplete readiness submission
      description: Escalate store readiness to a target team for resolution.
      operationId: createFieldReadinessEscalation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The store ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFieldReadinessEscalationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldReadinessEscalationRecord'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Store Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /readiness-escalations:
    get:
      tags:
        - Field Readiness
      summary: List readiness escalations
      description: Returns a list of escalations for control-panel operator queue.
      operationId: listFieldReadinessEscalations
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFieldReadinessEscalationsResponse'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /readiness-escalations/{id}:
    patch:
      tags:
        - Field Readiness
      summary: Update readiness escalation status
      description: Updates the status of a readiness escalation.
      operationId: updateFieldReadinessEscalation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The escalation ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFieldReadinessEscalationRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldReadinessEscalationRecord'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Escalation Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stores/{id}/readiness-approvals:
    post:
      tags:
        - Field Readiness
      summary: Approve or reject store readiness
      description: Performs a formal readiness approval or rejection.
      operationId: createFieldReadinessApproval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The store ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFieldReadinessApprovalRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldReadinessApprovalRecord'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Store Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stores/{id}/readiness-approvals/latest:
    get:
      tags:
        - Field Readiness
      summary: Get latest readiness approval decision
      description: Returns the latest approval or rejection record for a store.
      operationId: getLatestFieldReadinessApproval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The store ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldReadinessApprovalRecord'
        '404':
          description: No approval record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orders/{id}:
    get:
      tags:
        - Order Lifecycle
      summary: Get order details
      description: Returns the order record, items, status history events, and support tickets.
      security:
        - BearerAuth: []
      x-bthwani-auth-dev-note: DEV_ONLY — backend accepts X-Client-Id header; actor scoping remains blocked on production auth roles
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailsResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid client identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /support/escalations:
    post:
      tags:
        - Support Escalation
      summary: Create a support escalation ticket
      description: Creates a support escalation ticket for a specific order.
      operationId: createSupportEscalation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupportEscalationRequest'
      responses:
        '201':
          description: Support ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportEscalationRecord'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional bearer token for personalized store discovery (e.g. retrieving user's favorite stores).
  schemas:
...
    CreateFieldReadinessEscalationRequest:
      type: object
      required:
        - reason
        - target_team
      properties:
        field_agent_id:
          type: string
        reason:
          type: string
        target_team:
          type: string
          enum: [partner-management, control-panel, marketing]
    FieldReadinessEscalationRecord:
      type: object
      required:
        - id
        - store_id
        - reason
        - target_team
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        store_id:
          type: string
        field_agent_id:
          type: string
        reason:
          type: string
        target_team:
          type: string
          enum: [partner-management, control-panel, marketing]
        status:
          type: string
          enum: [escalated, info_requested, resolved, rejected]
        operator_note:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateFieldReadinessEscalationRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [info_requested, resolved, rejected]
        operator_note:
          type: string
    ListFieldReadinessEscalationsResponse:
      type: object
      required:
        - escalations
        - pagination
      properties:
        escalations:
          type: array
          items:
            $ref: '#/components/schemas/FieldReadinessEscalationRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateFieldReadinessApprovalRequest:
      type: object
      required:
        - decision
      properties:
        operator_id:
          type: string
        decision:
          type: string
          enum: [approved, rejected]
        reason:
          type: string
    FieldReadinessApprovalRecord:
      type: object
      required:
        - id
        - store_id
        - decision
        - created_at
      properties:
        id:
          type: string
        store_id:
          type: string
        operator_id:
          type: string
        decision:
          type: string
          enum: [approved, rejected]
        reason:
          type: string
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    # ─── J-002 / DSH-SLICE-002A: Product Identity ─────────────────────────────
...
    OrderRecord:
      type: object
      required:
        - id
        - store_id
        - client_id
        - status
        - total_price
        - created_at
        - updated_at
      properties:
        id:
          type: string
        store_id:
          type: string
        client_id:
          type: string
        status:
          type: string
          enum: [CREATED, ACCEPTED, READY_FOR_PICKUP, DELIVERED, CANCELLED, REFUNDED, ACCEPTED_BY_CAPTAIN, PICKED_UP, EN_ROUTE, ARRIVED]
        total_price:
          type: number
        wlt_payment_ref_id:
          type: string
        wlt_refund_ref_id:
          type: string
        captain_id:
          type: string
        captain_latitude:
          type: number
        captain_longitude:
          type: number
        captain_lifecycle_status:
          type: string
        pod_media_key:
          type: string
          nullable: true
          description: Proof-of-delivery media key reference (set on DELIVERED). DSH stores reference only, not the raw binary.
        delivery_failure_reason:
          type: string
          nullable: true
          description: Reason delivery failed (set on FAILED_DELIVERY or RETURNING_TO_STORE).
        wlt_refund_trigger_ref:
          type: string
          nullable: true
          description: WLT bridge reference ID. DSH stores only — WLT (004E) owns refund execution independently.
        wlt_settlement_ref_id:
          type: string
          nullable: true
          description: WLT settlement transaction reference.
        settlement_status:
          type: string
          enum: [NOT_SETTLED, SETTLEMENT_PENDING, SETTLED, SETTLEMENT_FAILED]
          default: NOT_SETTLED
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItemRecord:
      type: object
      required:
        - id
        - order_id
        - product_id
        - quantity
        - price
      properties:
        id:
          type: string
        order_id:
          type: string
        product_id:
          type: string
        quantity:
          type: integer
        price:
          type: number
    OrderStatusEventRecord:
      type: object
      required:
        - id
        - order_id
        - actor
        - from_status
        - to_status
        - created_at
      properties:
        id:
          type: string
        order_id:
          type: string
        actor:
          type: string
          enum: [client, partner, captain, operator, system]
        from_status:
          type: string
        to_status:
          type: string
        note:
          type: string
        created_at:
          type: string
          format: date-time
    SupportEscalationRecord:
      type: object
      required:
        - id
        - order_id
        - actor
        - issue_type
        - description
        - status
        - created_at
      properties:
        id:
          type: string
        order_id:
          type: string
        actor:
          type: string
          enum: [client, partner]
        issue_type:
          type: string
          enum: [delayed_delivery, wrong_items, missing_items, payment_issue, other]
        description:
          type: string
        status:
          type: string
          enum: [open, in-review, resolved]
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
...
    CreateSupportEscalationRequest:
      type: object
      required:
        - order_id
        - actor
        - issue_type
        - description
      properties:
        order_id:
          type: string
        actor:
          type: string
          enum: [client, partner]
        issue_type:
          type: string
          enum: [delayed_delivery, wrong_items, missing_items, payment_issue, other]
        description:
          type: string
    OrderDetailsResponse:
      type: object
      required:
        - order
        - items
        - status_events
        - support_tickets
      properties:
        order:
          $ref: '#/components/schemas/OrderRecord'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRecord'
        status_events:
          type: array
          items:
            $ref: '#/components/schemas/OrderStatusEventRecord'
        support_tickets:
          type: array
          items:
            $ref: '#/components/schemas/SupportEscalationRecord'
    Pagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

    # ─── J-003 / Checkout Schemas ─────────────────────────────────────────────
...